requirements.txt
django-filter,
xhtml2pdf (not added to settings.py file)